+++ src/keys/ecdsa.rs
@@ -165,7 +165,7 @@ impl PartialEq for EcDsaPublicKey {
 
 impl fmt::Display for EcDsaPublicKey {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.write_str(&serialize_ossh_pubkey(self, None).unwrap())
+        f.write_str(&serialize_ossh_pubkey(self, "").unwrap())
     }
 }